backend/pyproject.toml
[tool.poetry]
name = "vk-sporty-beavers-backend-2-0"
version = "0.1.0"
description = ""
authors = ["Mikhail Isaev <original.smile2002@gmail.com>"]
readme = "README.md"
packages = [{include = "vk_sporty_beavers_backend_2"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
httpx = "^0.23.3"
requests = "^2.28.2"
python-multipart = "^0.0.5"
python-dotenv = "^1.0.0"
python-decouple = "^3.7"
psycopg2-binary = "^2.9.5"
sqlalchemy = "^2.0.4"
gunicorn = "^20.1.0"
boto3 = "^1.26.94"
numpy = "^1.24.2"
opencv-python = "4.5.5.62"
celery = "^5.2.7"
pyjwt = "^2.6.0"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.0.1"
black = "^23.1.0"
flake8 = "^6.0.0"
mypy = "^1.0.1"
pylint = "^2.16.2"
pydantic = "^1.10.5"
types-requests = "^2.28.11.15"
isort = "^5.12.0"
unify = "^0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
